- document CETEIcean update process; - switch to one-file distribution of the CETEIcean JavaScript;
