+++ docs/about.md
@@ -154,7 +154,7 @@ Static site generated with [Jekyll](https://jekyllrb.com/) using customized
 
 Static site is hosted on [GitHub Pages](https://pages.github.com/).
 
-Facsimiles are hosted in a Google Cloud Storage bucket `facsimiles.alter-rebbe.org`
+Facsimiles are hosted in a public Google Cloud Storage bucket `facsimiles.alter-rebbe.org`
 and are retrievable only using exact URLs. 
 
 To rebuild collection indices and verify name references:
@@ -166,3 +166,16 @@ The site can be served locally by running the following command in `docs`:
 ```
   $ bundle exec jekyll serve
 ```
+
+To update CETEIcean to the current repository version:
+
+```
+  $ git clone git@github.com:TEIC/CETEIcean.git
+  $ cd CETEIcean
+  $ npm clean-install
+  $ npm run-script build
+```
+
+Merge into the archive project taking local changes into account:
+- `/test/CETEIcean.css` into `/docs/_sass/CETEIcean.scss`
+- `/dist/CETE.js` into /docs/js/CETEI.js`
